Remove "this email was deleted by another client"
I was on the fence about adding this in the first place. The state of an
email being \Deleted but still in this inbox is unusual, and unlikely to
occur unless the user is already somewhat knowledgable about IMAP and
utilizing power-user-level tooling which could cause the situation to
arise. Alps does not target that kind of user, so this can be hidden.