Fix form data loss after invalidated session
A session can become invalid without explicitly logging out due to the idle
timeout or a server restart. A user may submit a form after the session has
been invalidated: for example, the user may send a message or save a draft
after session expires.

Currently, we redirect using 302 Found to /login?next=original-url.
If a login token exists, we authenticate, skip the login form and redirect back
to the original URL. This results in the loss of the POST method and body of
the request. The body can be a composed message.

Preserve the method and body of the unauthenticated request using 307
Temporary Redirect instead.

This does not prevent data loss if there is no login token available. In this
case, the user gets redirected to the login form. After submission, the request
is a POST to /login containing username and password. Information about the
original URL and request body does not propagate.

Link: https://lists.sr.ht/~migadu/alps-devel/%3C0fc4ae1bdc94d3f036f4c8f0001e5205%40c3f.net%3E