login: add encrypted token with session lifetime
The current behavior is to preserve user sessions between application
restarts, only if the user has enabled 'remember me'. The lifetime of
the persistent login token is a month.

However, the rest of the sessions are lost between deployments. Users
will experience unexpected logout and even lose data submitted after
the shutdown. The goal is to avoid incorrect session invalidation.

This commit introduces a session login token with lifetime that
covers the duration of the session, but does not last long after the
session ends. The terms 'lifetime' and 'expiration', here, refer to the
time-to-live of the encrypted token. The cookie is not persistent as it
only needs to be available as long as the session cookie itself. The
session token has sliding expiration and short expiry, by default set to
the idle timeout. With an idle timeout of 30', the token remains
valid, in the worst case, for 30' after the user has logged out.

The session token is encrypted and contains the user credentials,
similarly to the 'remember me' token. It is enabled by default and used
alongside the session cookie and the optional 'remember me' token.