imap: handle multiple replies when getting message
It is possible that UID FETCH for a single message returns multiple data
responses. As reported in the issue, a Gmail server returns the
following when fetching an unseen message:

* 9 FETCH (UID 40973 FLAGS (\Seen) ENVELOPE (...) BODY[1] ...)
* 9 FETCH (UID 40973 FLAGS (\Seen))

As we expect a single response, we pass a buffered channel with
capacity one to the IMAP client. Sending the second response blocks.
Waiting for the FETCH operation to finish results in hanging.

Ensure sending to the result channel does not block indefinitely.
Keep the first response message and discard the rest.

Fixes: https://todo.sr.ht/~migadu/alps/156