caldav: fix mapping of events to dates
Truncating the starting time of an event can result in an incorrect
starting date. Given that truncate operates on time as an absolute
duration, it is possible that two events belonging to the same calendar
day get mapped to different dates.

Instead, use the date format of the time as the map key.