Upgrade go-webdav dependency
This commit upgrades go-webdav to what is currently the master branch.
The go-webdav client implementation gained some features with regards to
service discovery that alps can benefit from.

With this change, one can pass e.g. the following upstream URL:

   carddav+insecure://example.com/.well-known/carddav

This URL is defined in [RFC 6764, section 6][1]. If the server supports
it, it will respond with a redirect to the currently authenticated
user's principal URL. In the version used so far, the go-webdav client
does not handle this redirect well.

[1]: https://datatracker.ietf.org/doc/html/rfc6764#section-6

See also: https://github.com/emersion/go-webdav/commit/0363312