Fix proxying of images in HTML view
In Firefox parent cookies do not get sent in iframes using srcdoc [0].
As a result, proxying requests get redirected to the login page.

Implement a work-around until the respective browser bug is fixed.
Instead of embedding the sanitized HTML view in the iframe using srcdoc,
serve it separately and set src to the URL of the route.

[0]: https://bugzilla.mozilla.org/show_bug.cgi?id=1741489
References: https://todo.sr.ht/~migadu/alps/152