Add ManageSieve client
This parses the configured upstream ManageSieve server URL. If no configured
server matches, it attempts to resolve the hostname using an SRV lookup.

It connects to the server using an unencrypted connection. Then, it upgrades
to an encrypted one and authenticates using the PLAIN mechanism.

Implements: https://todo.sr.ht/~migadu/alps/9